V3.0.0 Alpha 10 (#61)
* More tournament work.

* More tournament work.

* A lot more tournament work.

* More tournament work. Generator command created. Other generators modified to reduce code duplication and increase test information generation.

* More tournament work. Split ActivityManager.

* v3.0.0 Alpha 9 Patch 14
* Fix time to grant timer not being rounded to a full second. (This went live a few days ago, I just realized I never committed it.)

* Revert "v3.0.0 Alpha 9 Patch 14"

This reverts commit 4ca055e497ad3b4a1086f2e8236358c4a2548174.

* Melee, non-FFA tournaments are almost ready to be tested.

* Tournament debugging. Need to rework how injuries are counted in tournaments though.

* Tournament debugging. Melee tournaments work. Need to work on report viewing.

* * Add realm restore command.
* Add realm statistics for messages and conversations.
* Stop showing heraldry of people that don't have any.
* Add tournament overall report viewing.
* Remove targets from fallen fighters.
* Remove fallen fighters as other's targets.

* * Finish tournament display for 1v1 in mastery.
* Tweak activity report display to be a bit more easily readable.
* Initial work on tournaments for legacy.

* Solos and Duos now work. Fixed bug in target selection (might need randomizing but check 5v5 first). Legacy and mastery both function for tournaments.

* All melee tournament types should work now. Join form works. Create form is restricted to just melees for this update. Added some more doctrine debug stuff. Added tournament running to GameRunner. Fixed typo in UpdateDatabaseCommand.

* Melee Tournaments should be complete now. Just need to add logic to open the gates.

* Should be the last of A10.

* Minor tweaks.

* Added character status updating.

* Add fishing to activities manual page.

+++ src/Command/ProcessMessagesCommand.php
@@ -3,12 +3,15 @@
 namespace App\Command;
 
 use App\Entity\DelayedMessage;
+use App\Service\CommonService;
 use App\Service\ConversationManager;
+use App\Service\NotificationManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\ErrorHandler\Exception\FlattenException;
 use Symfony\Component\Stopwatch\Stopwatch;
 
 
@@ -17,6 +20,8 @@ class ProcessMessagesCommand extends Command {
 	public function __construct(
 		private ConversationManager $conv,
 		private EntityManagerInterface $em,
+		private NotificationManager $nm,
+		private CommonService $cs,
 	) {
 		parent::__construct();
 	}
@@ -37,18 +42,54 @@ class ProcessMessagesCommand extends Command {
 			$stopwatch->start('messages');
 		}
 
-		$all = $this->em->getRepository(DelayedMessage::class)->findAll();
-		/** @var DelayedMessage $m */
-		$i = 0;
-		$j = 0;
-		foreach ($all as $m) {
-			$function = $m->getTarget();
-			$j += $this->conv->$function(false, $m->getSender(), $m->getTopic(), $m->getContent(), $m->getSystemContent(), $m->getType());
-			$i++;
-			$this->em->remove($m);
-			$this->em->flush();
+		if ($this->cs->getGlobal('actions.running') == 0) {
+			$all = $this->em->getRepository(DelayedMessage::class)->findAll();
+			if (count($all) > 0) {
+				$i = 0;
+				$j = 0;
+				$now = new \DateTime('now')->format('Y-m-d H:i:s');
+				$output->writeln($now . ' - Delayed Message processing running...');
+				$this->cs->setGlobal('messages.running', 1);
+				$this->cs->setGlobal('messages.last', $now);
+				$stopwatch = new Stopwatch();
+				$stopwatch->start('actions');
+				/** @var DelayedMessage $m */
+				$failed = false;
+				try {
+					foreach ($all as $m) {
+						$function = $m->getTarget();
+						$j += $this->conv->$function(false, $m->getSender(), $m->getTopic(), $m->getContent(), $m->getSystemContent(), $m->getType());
+						$i++;
+						$this->em->remove($m);
+						$this->em->flush();
+					}
+				} catch (\Exception $e) {
+					try {
+						$error = new FlattenException()::createFromThrowable($e);
+						$msg = $error->getMessage();
+						$code = $error->getCode();
+						$trace = $error->getTraceAsString();
+						$txt = "Status code: $code\nError: $msg\nTrace: $trace";
+						$output->writeln($txt);
+						$this->nm->spoolError($txt);
+						$failed = true;
+					} catch (\Exception $f) {
+						$output->writeln($f->getMessage());
+						$output->writeln($f->getTraceAsString());
+						$failed = true;
+					}
+				}
+				if (!$failed) {
+					$this->cs->setGlobal('messages.running', 0);
+					$now = new \DateTime('now')->format('Y-m-d H:i:s');
+					$this->cs->setGlobal('messages.last', $now);
+					$this->cs->setGlobal('messages.reported', 0);
+				}
+
+				$output->writeln("$i delayed messages sent resulting in $j new messages in game.");
+			}
+
 		}
-		$output->writeln("$i delayed messages sent resulting in $j new messages in game.");
 		$count = $this->em->createQuery('DELETE FROM App\Entity\ChatMessage m WHERE m.ts < :date')->setParameters(['date'=>new \DateTime("-14 days")])->execute();
 		$output->writeln("$count chat messages older than 14 days have been deleted.");
 		if ($timing) {
